backend/requirements.txt
flashrank
fastapi
uvicorn
openai
pinecone-client
python-dotenv
cohere
boto3
